dom/indexedDB/test/test_file_quota.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const READ_WRITE = IDBTransaction.READ_WRITE;

    const name = window.location.pathname;

    const objectStoreName = "Blobs";

    const fileData = { key: 1, file: getNullFile("random.bin", DEFAULT_QUOTA) };

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onupgradeneeded = grabEventAndContinueHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield;

    is(event.type, "upgradeneeded", "Got correct event type");

    let db = event.target.result;

    db.createObjectStore(objectStoreName, { });

    event = yield;

    is(event.type, "success", "Got correct event type");

    trans = db.transaction([objectStoreName], READ_WRITE);
    let objectStore = trans.objectStore(objectStoreName);

    request = objectStore.add(fileData.file, fileData.key);
    request.addEventListener("error", new ExpectError("UnknownError", true));
    request.onsuccess = unexpectedSuccessHandler;
    event = yield;

    trans.oncomplete = grabEventAndContinueHandler;
    event = yield;

    is(event.type, "complete", "Got correct event type");

    finishTest();
    yield;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest(true);"></body>

</html>
